stock_rag/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.339
openai==1.3.5
faiss-cpu==1.7.4
pandas==2.1.3
yfinance==0.2.31
python-dotenv==1.0.0
pydantic==2.5.2
elasticsearch==7.17.4
beautifulsoup4==4.12.2
requests==2.31.0
